+++ src/Exceptions/ArionumException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace pxgamer\Arionum\Exceptions;
+
+use RuntimeException;
+
+abstract class ArionumException extends RuntimeException
+{
+}