+++ src/Exceptions/GenericApiException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace pxgamer\Arionum\Exceptions;
+
+final class GenericApiException extends ArionumException
+{
+}